rename duplicity.py - to ensure not issues (duplicity's own library is also called duplicity)
